Create README.md
The Optimization_c.m file is the main program that runs three algorithms described in the paper :

IAG , ICAG and IG

The function Y_l.mat file contains the input to the sensors which are loaded. 

cluster_gradient.m calculates the gradient sum of a cluster of sensors. 

function_prototype_grad.m calculates the gradient for one sensor 

function_prototype.m contains the function but it is not needed 

objective_function evaluates the total objective function and it is not needed either.